ubox: run init script through shellcheck
Warnings fixed:

SC2004: $/${} is unnecessary on arithmetic variables.
SC2166: Prefer [ p ] && [ q ] as [ p -a q ] is not well defined.
SC2086: Double quote to prevent globbing and word splitting.

Removed most usages of ${} with $. There's no need for it. ${} is
mainly useful with substrings and arrays, which are not used here.

Signed-off-by: Rosen Penev <rosenp@gmail.com>