toolchain/musl: remove several GNU headers
Remove GLOB_ONLYDIR patch. Only fstools relies on it. fstools can be fixed
separately.

Remove woresize.h file. It seems to be for an old version of GCC.

Remove features.h and glibc-types files. Same as above.

Replace sys/cdefs.h with what is in alpine, openembedded, and yocto. The
actual header is deprecated and should not be used. This version is
smaller than the glibc version currently in the tree.

Explanation from buildroot:

  sys/cdefs.h is a non-standard header, originating from glibc,
  that defines non-standard and legacy macros. This header is not
  always available (e.g. musl does not provide it, but Buildroot
  installs a surrogate, minimalist one in this case).

  Some packages use this header, which is wrong, as that makes
  then non-portable.

  By default, Buildroot does not detect the use of that header.
  In the future, this default may change to warning, or even to
  erroring out.

Tested with all packages in the base tree. They all compile.

Signed-off-by: Rosen Penev <rosenp@gmail.com>