lantiq: 2812hnu-f3: : embed EEPROM in dts
Apparently to get wifi working on this unit an external EEPROM that
seems to not be available on flash is needed. Instead of asking users to
hunt for it, just embed it in dts.

Data is taken from the P2812HNUFx-Pre-Build github repository where it is
present as the file:

RT3092.eeprom

Signed-off-by: Rosen Penev <rosenp@gmail.com>